Entries löschbar machen

+++ src/app/components/value-modal/value-modal.component.ts
@@ -1,5 +1,6 @@
-import { Component, Input, OnChanges, SimpleChanges } from "@angular/core";
+import { Component, EventEmitter, Input, OnChanges, Output, SimpleChanges } from "@angular/core";
 import { ToastService } from "angular-toastify";
+import * as bootstrap from "bootstrap";
 import { PortfolioService } from "src/app/services/portfolio.service";
 import { PortfolioEntry } from "src/app/shared/interfaces/PortfolioEntry";
 
@@ -12,6 +13,7 @@ export class ValueModalComponent implements OnChanges {
 
   constructor(private portfolioService: PortfolioService, private _toastService: ToastService) { }
 
+  @Output() valueChanged = new EventEmitter<boolean>();
   @Input() portfolioId: number | null = null;
   @Input() entryId: number | null = null;
   entry: PortfolioEntry | null = null;
@@ -47,25 +49,34 @@ export class ValueModalComponent implements OnChanges {
     this.addValueActive = !this.addValueActive;
   }
 
+  // Zeigt Modal
+  show() {
+    const modal = new bootstrap.Modal(document.getElementById("valueModal") as Element);
+    modal.show();
+  }
+
   // Löscht entryvalue
   deleteValue(id: number) {
     this.error.msg = "";
     if (this.portfolioId != null && this.entryId != null) {
-      this.portfolioService.deletePortfolioEntryValue(this.portfolioId, this.entryId, id).subscribe({
-        next: (response) => {
-          if (response.success) {
-            this._toastService.success("Eintrag erfolgreich entfernt!");
-            this.getEntry();
-          } else {
-            this.error = response.error;
+      if (confirm("Soll dieser Wert wirklich entfernt werden?")) {
+        this.portfolioService.deletePortfolioEntryValue(this.portfolioId, this.entryId, id).subscribe({
+          next: (response) => {
+            if (response.success) {
+              this._toastService.success("Eintrag erfolgreich entfernt!");
+              this.getEntry();
+              this.valueChanged.emit(true);
+            } else {
+              this.error = response.error;
+            }
+          },
+          error: (error) => {
+            const errorTxt = "Beim Löschen des Eintrags ist ein Fehler aufgetreten!"
+            this.error.msg = errorTxt;
+            console.error(errorTxt, error);
           }
-        },
-        error: (error) => {
-          const errorTxt = "Beim Löschen des Eintrags ist ein Fehler aufgetreten!"
-          this.error.msg = errorTxt;
-          console.error(errorTxt, error);
-        }
-      })
+        })
+      }
     }
   }
